requirements.txt
flask==2.0.1
werkzeug==2.0.2
flask-restful==0.3.9
pymongo==4.1.1
marshmallow==3.13.0
flask-marshmallow==0.14.0
flask-cors==3.0.10
python-dotenv==0.19.0
bcrypt==3.2.0
pydantic==1.8.2
email-validator==1.1.3